+++ package.json
@@ -19,10 +19,10 @@
   "dependencies": {
     "@lottiefiles/react-lottie-player": "3.5.4",
     "@nextui-org/react": "2.4.2",
-    "@polkadot/api": "11.2.1",
+    "@polkadot/api": "11.3.1",
     "@polkadot/keyring": "12.6.2",
     "@polkadot/react-identicon": "3.6.6",
-    "@polkadot/types": "11.2.1",
+    "@polkadot/types": "11.3.1",
     "@polkadot/util": "12.6.2",
     "@polkadot/util-crypto": "12.6.2",
     "@remix-run/node": "2.9.2",
@@ -61,7 +61,7 @@
     "@vitest/coverage-v8": "1.6.0",
     "autoprefixer": "10.4.19",
     "concurrently": "8.2.2",
-    "cssnano": "7.0.2",
+    "cssnano": "7.0.3",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-import-resolver-alias": "1.1.2",
@@ -71,7 +71,7 @@
     "eslint-plugin-import": "2.29.1",
     "eslint-plugin-json": "4.0.0",
     "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-react": "7.34.2",
+    "eslint-plugin-react": "7.34.3",
     "express-csp-header": "^5.2.1",
     "husky": "9.0.11",
     "jsdom": "24.1.0",
@@ -81,6 +81,7 @@
     "tailwind-merge": "2.3.0",
     "tailwindcss": "3.4.4",
     "typescript": "5.5.2",
+    "typescript-remix-routes-plugin": "^1.0.1",
     "vite": "5.3.1",
     "vite-plugin-compression2": "^1.1.1",
     "vite-plugin-svgr": "4.2.0",
