--- src/app/routes/receive/index.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { ReceivePage } from '@/screens/receive';
-
-const Page: FC = () => {
-  return <ReceivePage />;
-};
-
-export default Page;
