--- src/app/routes/receive/select-token.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { ReceiveTokenPage } from '@/screens/receive';
-
-const Page: FC = () => {
-  return <ReceiveTokenPage />;
-};
-
-export default Page;
