--- src/app/routes/transfer/direct/address.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { AddressPage } from '@/screens/transfer';
-
-const Page: FC = () => {
-  return <AddressPage />;
-};
-
-export default Page;
