--- src/app/routes/transfer/direct/amount.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { AmountPage } from '@/screens/transfer';
-
-const Page: FC = () => {
-  return <AmountPage />;
-};
-
-export default Page;
