--- src/app/routes/transfer/direct/confirmation.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { ConfirmationPage } from '@/screens/transfer';
-
-const Page: FC = () => {
-  return <ConfirmationPage />;
-};
-
-export default Page;
