--- src/app/routes/transfer/direct/result.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { ResultPage } from '@/screens/transfer';
-
-const Page: FC = () => {
-  return <ResultPage />;
-};
-
-export default Page;
