--- src/app/routes/transfer/gift/amount.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { AmountGiftPage } from '@/screens/transfer';
-
-const Page: FC = () => {
-  return <AmountGiftPage />;
-};
-
-export default Page;
