--- src/app/routes/transfer/select-token.tsx
@@ -1,9 +1,0 @@
-import { type FC } from 'react';
-
-import { SelectTokenPage } from '@/screens/transfer';
-
-const Page: FC = () => {
-  return <SelectTokenPage />;
-};
-
-export default Page;
