+++ src/components/Assets/AssetBalance.tsx
@@ -1,3 +1,5 @@
+import { type FC } from 'react';
+
 import { type IconNames } from '../Icon/types';
 
 import { type Asset } from '@/common/chainRegistry/types';
@@ -16,7 +18,7 @@ type Props = {
   animate?: boolean;
 };
 
-const AssetBalance = ({ balance, asset, name, className, showPrice, showArrow, animate }: Props) => {
+const AssetBalance: FC<Props> = ({ balance, asset, name, className, showPrice, showArrow, animate }) => {
   const { precision, symbol, priceId } = asset;
 
   return (
