--- src/screens/dashboard/index.ts
@@ -1,1 +1,0 @@
-export { DashboardMain } from './DashboardMain';
