--- src/screens/exchange/index.ts
@@ -1,5 +1,0 @@
-import ExchangePage from './ExchangePage';
-import MercuryoWidgetPage from './MercuryoWidgetPage';
-import SelectTokenExchangePage from './SelectTokenExchangePage';
-
-export { SelectTokenExchangePage, MercuryoWidgetPage, ExchangePage };
