--- src/screens/receive/index.ts
@@ -1,4 +1,0 @@
-import ReceivePage from './ReceivePage';
-import ReceiveTokenPage from './ReceiveTokenPage';
-
-export { ReceivePage, ReceiveTokenPage };
