--- src/screens/transfer/index.ts
@@ -1,19 +1,0 @@
-import AddressPage from './AddressPage';
-import ConfirmationPage from './ConfirmationPage';
-import CreateGiftPage from './CreateGiftPage';
-import ResultPage from './ResultPage';
-import SelectTokenPage from './SelectTokenPage';
-import TransferPage from './TrasferPage';
-import AmountGiftPage from './amount/AmountGiftPage';
-import AmountPage from './amount/AmountPage';
-
-export {
-  TransferPage,
-  SelectTokenPage,
-  AddressPage,
-  AmountPage,
-  AmountGiftPage,
-  ConfirmationPage,
-  ResultPage,
-  CreateGiftPage,
-};
